% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_spikes.R
\name{detect_spikes}
\alias{detect_spikes}
\title{Estimate spike train, underlying calcium concentration, and changepoints based on 
fluorescence trace for a range of tuning parameters.}
\usage{
detect_spikes(dat, gam, constraint, EPS, lambda_min, lambda_max,
  sensitivity = 1)
}
\arguments{
\item{dat}{fluorescence data}

\item{gam}{a scalar value for the AR(1) decay parameter.}

\item{constraint}{boolean specifying whether constrained or unconstrained optimization
problem}

\item{EPS}{double specfying the minimum calcium value}

\item{min_lambda}{smallest tuning parameter}

\item{max_lambda}{largest tuning parameter}
}
\value{
Returns a list with elements:

\code{spikes} the set of spikes

\code{fittedValues} estimated calcium concentration

\code{changePts} the set of changepoints

\code{cost} the cost at each time point (vector)

\code{nIntervals} the number of intervals at each point (vector)
}
\description{
Given [lambda_min, lambda_max] this procedure selects lambdas such that each
lambda is associated with a different number of spike events.
}
\details{
This algorithm solves the optimization problems

 \strong{AR(1) model:}
 
 minimize_{c1,...,cT} 0.5 sum_{t=1}^T ( y_t - c_t )^2 + lambda sum_{t=2}^T 1_[c_t != max(gam c_{t-1}, EPS)]
 
 for the global optimum, where y_t is the observed fluorescence at the tth timepoint.

\strong{Constrained AR(1) model:}

minimize_{c1,...,cT} 0.5 sum_{t=1}^T ( y_t - c_t )^2 + lambda sum_{t=2}^T 1_[c_t != max(gam c_{t-1}, EPS)]

c_t >= max(gam c_{t-1}, EPS), t = 2, ..., T

For additional information see: 

1. Jewell, Hocking, Fearnhead, and Witten (2018) <arXiv:1802.07380> and 

2. Jewell and Witten (2017) <arXiv:1703.08644>
}
\examples{

library(LZeroSpikeInference)
sim <- simulateAR1(n = 500, gam = 0.95, poisMean = 0.009, sd = 0.05, seed = 1)
plot(sim)

## Fits for a single tuning parameter

# AR(1) model
fit <- ARFPOP(dat = sim$fl, gam = 0.95, lambda = 1)
print(fit)

# compute fitted values from prev. fit
fit <- fit_ar1_model(fit)
plot(fit)

# or 
fit <- ARFPOP(dat = sim$fl, gam = 0.95, lambda = 1, compute_fitted_values = T)
plot(fit)

# Constrained AR(1) model
fit <- ARFPOP(dat = sim$fl, gam = 0.95, lambda = 1, constraint = T, compute_fitted_values = T)
print(fit)
plot(fit)

}
\seealso{
\strong{Estimate spikes:}
\code{\link{ARFPOP}}
\code{\link{fit_ar1_model}}
}
