\name{H3K4me3_PGP_immune_chunk24}
\alias{H3K4me3_PGP_immune_chunk24}
\docType{data}
\title{
  H3K4me3_PGP_immune ChIP-seq data set chunk 24
}
\description{
  This data set revealed a bug in the PeakSegPDPA solver: at one point
  it recovered a less likely model than PeakSegDP for McGill0091, 13 segments.
}
\usage{data("H3K4me3_PGP_immune_chunk24")}
\format{
  A data frame with 66713 observations on the following 5 variables.
}
\source{
  http://cbio.mines-paristech.fr/~thocking/chip-seq-chunk-db/
}
\keyword{datasets}
