\name{H3K4me3_XJ_immune_chunk1}
\alias{H3K4me3_XJ_immune_chunk1}
\docType{data}
\title{
 H3K4me3_XJ_immune chunk 1
}
\description{
  Some test data sets for the Poisson PeakSeg Segment Neighborhood Algo.
}
\usage{data("H3K4me3_XJ_immune_chunk1")}
\format{
  A data frame with 18027 observations on 5 variables: cell.type,
  sample.id, chromStart, chromEnd, coverage. 
}
\keyword{datasets}
