\name{mclapply.or.stop}
\alias{mclapply.or.stop}
\title{mclapply or stop}
\description{Run mclapply but stop with an error instead of silently returning
a try-error.}
\usage{mclapply.or.stop(...)}
\arguments{
  \item{\dots}{passed to parallel::mclapply}
}

\value{List of values, same as lapply.}

\author{Toby Dylan Hocking}





