\name{oracleModelComplexity}
\alias{oracleModelComplexity}
\title{oracleModelComplexity}
\description{Compute Oracle model complexity from paper of Cleynen et al.}
\usage{oracleModelComplexity(bases, segments)}
\arguments{
  \item{bases}{
}
  \item{segments}{
}
}

\value{numeric vector of model complexity values.}

\author{Toby Dylan Hocking}





