% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.estimated_spike_paths}
\alias{plot.estimated_spike_paths}
\title{Plot number of spikes vs. tuning parameter}
\usage{
\method{plot}{estimated_spike_paths}(x, xlims = NULL, ...)
}
\arguments{
\item{x}{output from running estimate_spike_paths}

\item{xlims}{optional parameter to specify the x-axis limits}

\item{...}{arguments to be passed to methods}
}
\description{
Plot number of spikes vs. tuning parameter
}
\seealso{
\code{\link{estimate_spike_paths}},
\code{\link{estimate_spikes}},
\code{\link{estimate_calcium}},
}
