% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.estimated_spikes}
\alias{plot.estimated_spikes}
\title{Plot the solution to an L0 segmentation problem}
\usage{
\method{plot}{estimated_spikes}(x, xlims = NULL, ...)
}
\arguments{
\item{x}{output from running estimatedSpikes}

\item{xlims}{optional parameter to specify the x-axis limits}

\item{...}{arguments to be passed to methods}
}
\description{
Plot the solution to an L0 segmentation problem
}
\seealso{
\code{\link{estimate_spikes}},
\code{\link{estimate_calcium}},
}
