% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.estimated_spike_paths}
\alias{print.estimated_spike_paths}
\title{Print estimated spike path}
\usage{
\method{print}{estimated_spike_paths}(x, ...)
}
\arguments{
\item{x}{estimated spikes path}

\item{...}{arguments to be passed to methods}
}
\description{
Print estimated spike path
}
