% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.estimated_spikes}
\alias{print.estimated_spikes}
\title{Print estimated spikes}
\usage{
\method{print}{estimated_spikes}(x, ...)
}
\arguments{
\item{x}{estimated spikes}

\item{...}{arguments to be passed to methods}
}
\description{
Print estimated spikes
}
