\name{problem.predict.allSamples}
\alias{problem.predict.allSamples}
\title{problem predict allSamples}
\description{Predict for all samples in parallel.}
\usage{problem.predict.allSamples(set.dir, problem.name)}
\arguments{
  \item{set.dir}{project directory.}
  \item{problem.name}{name of problem to predict.}
}

\value{Nothing.}

\author{Toby Dylan Hocking}





