\name{problem.predict}
\alias{problem.predict}
\title{problem predict}
\description{Predict peaks for a genomic segmentation problem.}
\usage{problem.predict(problem.dir, model.RData)}
\arguments{
  \item{problem.dir}{Problem directory.}
  \item{model.RData}{Model file created via train_model.R}
}

\value{data.table of peak predictions.}

\author{Toby Dylan Hocking}





