\name{oracleModelComplexity}
\alias{oracleModelComplexity}
\title{oracleModelComplexity}
\description{Compute Oracle model complexity from paper of Cleynen et al.}
\usage{oracleModelComplexity(bases, 
    segments)}
\arguments{
  \item{bases}{bases }
  \item{segments}{segments }
}

\value{numeric vector of model complexity values.}

\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre]}





