\name{PeakSegPDPAchrom}
\alias{PeakSegPDPAchrom}
\title{PeakSegPDPAchrom}
\description{Helper function which returns a list of data.frames}
\usage{PeakSegPDPAchrom(input.dt, max.segments = NULL)}
\arguments{
  \item{input.dt}{data.table with columns count, chromStart, chromEnd.}
  \item{max.segments}{maximum number of segments, or NULL which means to keep going
until an active equality constraint is found.}
}



\author{Toby Dylan Hocking}





