% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.estimated_spikes}
\alias{plot.estimated_spikes}
\alias{plot.estimated_spikes}
\title{Plot the solution to an L0 segmentation problem}
\usage{
\method{plot}{estimated_spikes}(fits, xlims = NULL, ...)

\method{plot}{estimated_spikes}(fits, xlims = NULL, ...)
}
\arguments{
\item{fits}{output from running estimate_spike_paths}

\item{xlims}{optional parameter to specify the x-axis limits}

\item{...}{arguments to be passed to methods}

\item{x}{output from running estimate_spikes}

\item{xlims}{optional parameter to specify the x-axis limits}

\item{...}{arguments to be passed to methods}
}
\description{
Plot the solution to an L0 segmentation problem

Plot number of spikes vs. tuning paramter
}
\seealso{
\code{\link{estimate_spikes}},
\code{\link{estimate_calcium}},

\code{\link{estimate_spike_paths}},
\code{\link{estimate_spikes}},
\code{\link{estimate_calcium}},
}
