\name{problem.predict}
\alias{problem.predict}
\title{problem predict}
\description{Predict peaks for a genomic segmentation problem.}
\usage{problem.predict(problem.dir)}
\arguments{
  \item{problem.dir}{Problem directory.}
}

\value{data.table of peak predictions.}

\author{Toby Dylan Hocking}





